🚧 chore: app layout (in progress)

+++ src/external/components/line-chart/line-chart.go
@@ -25,5 +25,5 @@ type Dataset struct {
 	Data        []float64 `json:"data"`            // Points
 	Fill        bool      `json:"fill"`            // Remplissage des points
 	BorderColor string    `json:"backgroundColor"` // Couleur du tracé
-	Tension     float32   `json:"tension"`         // 0 -> droites entre les points, 1 -> courbes
+	Tension     float64   `json:"tension"`         // 0 -> droites entre les points, 1 -> courbes
 }
